Extract duplicate `HTTPClientSpy` into a shared helper folder
